feat(core): switch to use tasks-runner-v2 by default
