+++ scripts/e2e-ci2.sh
@@ -15,5 +15,7 @@ jest --maxWorkers=1 ./build/e2e/migrate.test.js &&
 jest --maxWorkers=1 ./build/e2e/move.angular.test.js &&
 jest --maxWorkers=1 ./build/e2e/move.workspace.test.js &&
 jest --maxWorkers=1 ./build/e2e/next.test.js &&
+jest --maxWorkers=1 ./build/e2e/nx-plugin.test.js &&
+jest --maxWorkers=1 ./build/e2e/print-affected.test.js &&
 jest --maxWorkers=1 ./build/e2e/delegate-to-cli.test.js &&
 jest --maxWorkers=1 ./build/e2e/downgrade-module.test.js
