+++ scripts/e2e-ci4.sh
@@ -9,7 +9,5 @@ mkdir -p tmp/nx
 export SELECTED_CLI=$1
 jest --maxWorkers=1 ./build/e2e/ng-add.test.js &&
 jest --maxWorkers=1 ./build/e2e/node.test.js &&
-jest --maxWorkers=1 ./build/e2e/nx-plugin.test.js &&
-jest --maxWorkers=1 ./build/e2e/print-affected.test.js &&
 jest --maxWorkers=1 ./build/e2e/react.test.js &&
 jest --maxWorkers=1 ./build/e2e/report.test.js
